Groups database operations moved into own file.
