+++ backend/offle_assistant/database/__init__.py
@@ -6,6 +6,7 @@ from ._users import (
     update_user_in_db,
     create_user_in_db,
 )
+
 from ._personas import (
     create_persona_in_db,
     update_persona_in_db,
@@ -13,15 +14,23 @@ from ._personas import (
     get_persona_by_id,
     delete_persona_by_id
 )
+
+from ._groups import (
+    get_group_by_id,
+    get_group_by_name,
+    get_default_group,
+    create_group,
+    delete_group,
+    update_group,
+)
+
 from ._queries import (
     get_message_history_entry_by_id,
     get_message_history_list_by_user_id,
     get_admin_exists,
-    get_group_by_id,
-    get_group_by_name,
-    get_default_group,
     get_message_history_entry_without_message_chain
 )
+
 from ._crud import (
     create_message_history_entry_in_db,
     update_message_history_entry_in_db,
@@ -29,38 +38,35 @@ from ._crud import (
     upload_file,
     download_file,
     delete_file,
-    create_group,
-    delete_group,
-    update_group
 )
 
 __all__ = [
-    "get_personas_by_creator_id",
+    "create_user_in_db",
     "get_user_by_email",
     "get_user_by_id",
+    "get_admin_exists",
+    "update_user_in_db",
+    "update_user_role_in_db",
+    "delete_user_in_db",
     "create_persona_in_db",
     "get_persona_by_id",
+    "get_personas_by_creator_id",
     "update_persona_in_db",
-    "delete_user_in_db",
-    "update_user_role_in_db",
-    "create_user_in_db",
-    "get_message_history_entry_by_id",
+    "delete_persona_by_id",
     "create_message_history_entry_in_db",
-    "update_message_history_entry_in_db",
+    "get_message_history_entry_by_id",
     "get_message_history_list_by_user_id",
-    "get_admin_exists",
+    "get_message_history_entry_without_message_chain",
+    "update_message_history_entry_in_db",
     "append_message_to_message_history_entry_in_db",
-    "update_user_in_db",
     "upload_file",
     "download_file",
     "delete_file",
     "create_group",
     "get_group",
-    "update_group",
     "get_group_by_id",
     "get_group_by_name",
-    "get_message_history_entry_without_message_chain",
     "get_default_group",
+    "update_group",
     "delete_group",
-    "delete_persona_by_id"
 ]
