+++ backend/offle_assistant/database/_queries.py
@@ -18,49 +18,6 @@ async def get_admin_exists(db: AsyncIOMotorDatabase) -> bool:
         return False
 
 
-DEFAULT_GROUP_NAME = "default"
-
-
-async def get_default_group(db: AsyncIOMotorDatabase):
-    """Fetch the default group, or create it if it doesn't exist."""
-    group = await db.groups.find_one({"name": DEFAULT_GROUP_NAME})
-
-    if not group:
-        # Create default group if it doesn't exist
-        group_data = {
-            "name": DEFAULT_GROUP_NAME,
-            "description": "Default group for all users"
-        }
-        insert_result = await db.groups.insert_one(group_data)
-        group = {**group_data, "_id": insert_result.inserted_id}
-
-    return group
-
-
-async def get_group_by_id(
-    group_id: str,
-    db: AsyncIOMotorDatabase
-) -> Optional[Dict]:
-    """Fetch a user from the database by their _id."""
-    return await db.groups.find_one({"_id": ObjectId(group_id)})
-
-
-async def get_group_by_name(
-    group_name: str,
-    db: AsyncIOMotorDatabase
-) -> Optional[Dict]:
-    """Fetch a user from the database by their _id."""
-    return await db.groups.find_one({"name": group_name})
-
-
-async def get_persona_by_id(
-    persona_id: str,
-    db: AsyncIOMotorDatabase
-) -> Optional[Dict]:
-    """Fetch a persona from the database by their _id."""
-    return await db.personas.find_one({"_id": ObjectId(persona_id)})
-
-
 async def get_message_history_entry_by_id(
     message_history_id: str,
     db: AsyncIOMotorDatabase
@@ -129,34 +86,6 @@ async def get_message_history_list_by_user_id(
     ).get(persona_id, [])
 
     return message_history
-
-
-async def get_personas_by_creator_id(
-    user_id: str,
-    db: AsyncIOMotorDatabase
-) -> Dict[str, str]:
-    """
-        Fetch all personas owned by a given user
-        and return as {persona_id: persona_name}.
-    """
-
-    try:
-        user_id = ObjectId(user_id)
-    except Exception as e:
-        logging.error("%s", e)
-        raise ValueError(f"Invalid user_id format: {user_id}")
-
-    # Find all personas where the creator_id matches the user's _id
-    personas = await db.personas.find(
-        {"creator_id": ObjectId(user_id)}
-    ).to_list(None)
-
-    # Convert the result to a dictionary {persona_id: persona_name}
-    persona_dict = {
-        str(persona["_id"]): persona["name"] for persona in personas
-    }
-
-    return persona_dict
 
 
 async def get_file_metadata(
